+++ CHANGES.rst
@@ -17,6 +17,8 @@ Changelog
 - using `seppl.variables` now
 - `fast-line-detection` and `hough-lines-prob` now store the length of the line segments in the object's
   meta-data as well
+- added `merge-segmentation-layers` filter to merging the layer of image segmentation item in storage with
+  the one currently passing through filter
 
 
 0.1.0 (2025-10-31)